% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_stats_utils.R
\name{calculate_beta_chisq}
\alias{calculate_beta_chisq}
\title{Function to calculate beta from standard error and p-value using Chi-square distribution}
\usage{
calculate_beta_chisq(se, p_value)
}
\description{
Function to calculate beta from standard error and p-value using Chi-square distribution
}
\keyword{internal}
