% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_stats_utils.R
\name{calculate_beta_normal}
\alias{calculate_beta_normal}
\title{Function to calculate beta from standard error and p-value using Normal distribution}
\usage{
calculate_beta_normal(se, p_value)
}
\description{
Function to calculate beta from standard error and p-value using Normal distribution
}
\keyword{internal}
