% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_stats_utils.R
\name{calculate_pvalue_chisq}
\alias{calculate_pvalue_chisq}
\title{Function to calculate p-value using the Chi-square distribution (df = 1)}
\usage{
calculate_pvalue_chisq(beta, se)
}
\description{
Function to calculate p-value using the Chi-square distribution (df = 1)
}
\keyword{internal}
