% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_stats_utils.R
\name{calculate_pvalue_normal}
\alias{calculate_pvalue_normal}
\title{Function to calculate p-value using the normal distribution (two-tailed)}
\usage{
calculate_pvalue_normal(beta, se)
}
\description{
Function to calculate p-value using the normal distribution (two-tailed)
}
\keyword{internal}
