% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_stats_utils.R
\name{calculate_se_normal}
\alias{calculate_se_normal}
\title{Function to calculate standard error from beta and p-value using Normal distribution}
\usage{
calculate_se_normal(beta, p_value)
}
\description{
Function to calculate standard error from beta and p-value using Normal distribution
}
\keyword{internal}
