% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{sg._copy_plink_win}
\alias{sg._copy_plink_win}
\title{Copy PLINK binary and reference panel into package directory (Windows)}
\usage{
sg._copy_plink_win(local_plink_dir)
}
\arguments{
\item{local_plink_dir}{A string. The full path to the local folder containing `plink.exe`
and any reference panel files (e.g., `.bed`, `.bim`, `.fam`) you wish to use.}
}
\value{
Invisibly returns the path to the destination folder where files were copied.
}
\description{
This function copies all files (including `plink.exe`, `.bed`, `.bim`, `.fam`, etc.)
from a specified local directory into the internal `plink/` folder inside the installed
package directory (i.e., where `system.file(package = "sequoiar")` points to).
}
