% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{sg.clump}
\alias{sg.clump}
\title{Local LD clumping using PLINK and EUR reference panel}
\usage{
sg.clump(df, clump_r2)
}
\arguments{
\item{df}{A data frame containing at least `SNP` and `mlogp` columns. `mlogp` must not contain Inf.}

\item{clump_r2}{Numeric. The r² threshold for LD clumping (e.g., 0.1).}
}
\value{
A data frame of SNPs that remain after clumping, with the same structure as input `df`.
}
\description{
This function performs LD clumping on a data frame using PLINK, based on SNP-level
`-log10(p-value)` significance and a provided LD threshold (`r2`). It expects PLINK
binary and a EUR reference panel to be placed in the package's `plink/` directory.
}
