% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{sg.clump_full}
\alias{sg.clump_full}
\title{Perform LD Clumping using PLINK}
\usage{
sg.clump_full(df, clump_r2, p_threshold = 0.05)
}
\arguments{
\item{df}{A data frame containing at least `SNP` and `mlogp` columns. `mlogp` must not contain Inf.}

\item{clump_r2}{Numeric. The r² threshold for LD clumping (e.g., 0.1).}

\item{p_threshold}{Numeric. The P-value threshold for filtering SNPs before clumping (default 0.05).}
}
\value{
A data frame of SNPs that remain after clumping, with the same structure as input `df`.
}
\description{
This function filters SNPs based on Linkage Disequilibrium (LD) using an
external PLINK binary and a reference panel. It writes temporary files,
runs PLINK, and returns the pruned dataframe.
}
