% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_LD_local.R
\name{sg.link_LD_local}
\alias{sg.link_LD_local}
\title{Append LD values to SNPs using local PLINK via `ieugwasr`}
\usage{
sg.link_LD_local(loc_data, lead_SNP)
}
\arguments{
\item{loc_data}{A locus object when making locus plot containing a `data` data frame with a `SNP` column.}

\item{lead_SNP}{A character string specifying the SNP to be used as the LD reference.
Must exist in `loc_data$data$SNP`.}
}
\value{
A modified version of `loc_data` with an added `ld` column in `loc_data$data`.
}
\description{
This function uses `ieugwasr::ld_matrix_local()` to compute LD (`r²`) between a specified lead SNP
and other SNPs in a given locus. It appends the LD values to the `data` frame inside the `loc_data` list.
It assumes a PLINK binary and EUR reference panel are available in the installed package directory.
}
